app_django/requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-rest-auth==1.1.0
Django==3.1.4
django-allauth==0.42.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-extensions==3.0.3
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
python3-openid==3.2.0
pytz==2020.5
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1

